use builtin `locate` during margin-note (closes #8)
Historical `margin-note` implementation shifted the margin note from where the function was called, so a reference label was needed as the position to avoid diverging layouts. However, there are some error cases where `query`ing an element yields the wrong location compared to using `locate` itself. So, the tradeoffs, esp. with newer versions of typst, are in favor of using `locate` directly during margin notes.