Don't number `note-outline` heading
This changes the default behavior of `note-outline` so that it doesn't
number the heading it generates.  The heading shouldn't be numbered by
default because typically this notes section isn't a part of the final
published document and would change the numbers of sections that are,
which can cause mistakes, e.g. writing the literal "Section 6" instead
of "Section 5".

Closes https://github.com/ntjess/typst-drafting/issues/36